face-recognition/requirements.txt
# Flask Web Framework
Flask==3.0.0
Flask-CORS==4.0.0
python-dotenv==1.0.0

# Face Recognition
face_recognition==1.3.0
dlib==19.24.0
numpy==1.24.3
Pillow==10.1.0

# HTTP Requests
requests==2.31.0

# Utilities
gunicorn==21.2.0